% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp.R
\name{emp}
\alias{emp}
\title{Generate empiricial test scripts for batch jobs}
\usage{
emp(type, path, taxo.num, loop.num)
}
\arguments{
\item{type}{a character object. Either "TCGA" or "ORIEN"}

\item{path}{a character object specifying where the scripts and RDS will be save to. Don't add "/" at the end}

\item{taxo.num}{an integer to select how many taxonomy to be tested on}

\item{loop.num}{an integer to set number of cycles for bootstrapping}
}
\value{
a R script and a sl script which can then be submitted as a job
}
\description{
Generate empiricial test scripts for batch jobs
}
