% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exoToDF.R
\name{exoToDF}
\alias{exoToDF}
\title{change data frame into wide format}
\usage{
exoToDF(
  data,
  surv,
  taxalevels = c("domain", "kingdom", "phylum", "class", "order", "family", "genus",
    "species")
)
}
\arguments{
\item{data}{input data in long format}

\item{surv}{survival information for the data}

\item{taxalevels}{character vector}
}
\value{

}
\description{
change data frame into wide format
}
