% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_surv_input.R
\name{generate_surv_input}
\alias{generate_surv_input}
\title{Title}
\usage{
generate_surv_input(cancer.subtype, input.list, tresh.list, area, tax_num)
}
\arguments{
\item{cancer.subtype}{a character used to select data from input list, tresh.list, area}

\item{input.list}{a list containing survival and taxonomy information}

\item{tresh.list}{a list containing threshold information}

\item{area}{a list containing area information derived from threshold data. It's used
to select taxonomy level with top area}

\item{tax_num}{a numeric value specifyingthe amount of taxonomy level to be selected}
}
\value{
a list conatining survival data, taxonomy data, threshold data, and selected taxonomy.
The list should be used as an input for plotting survival analysis
}
\description{
Title
}
