% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_emp.R
\name{plot_emp}
\alias{plot_emp}
\title{plot empirical test object and output statistics}
\usage{
plot_emp(list, plot.title)
}
\arguments{
\item{list}{a list object generated by function "emp"}

\item{plot.title}{a character object specifying the plot title}
}
\value{
a ggplot2 object and a dataframe
}
\description{
plot empirical test object and output statistics
}
