% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{plot_survival}
\alias{plot_survival}
\title{Create survival plot}
\usage{
plot_survival(modified_input, taxlev, data.type)
}
\arguments{
\item{modified_input}{an output list from function, containing input data for survival plot.}

\item{taxlev}{a character to select a taxonomy level to plot}

\item{data.type}{a character specifying where the data comes from and what is the cancer subtype. This
will be added to the title}
}
\value{
a survival plot
}
\description{
Create survival plot
}
