% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.area.R
\name{rank.area}
\alias{rank.area}
\title{Calarea}
\usage{
rank.area(taxlev, ptable.df, alpha = 0.05)
}
\arguments{
\item{taxlev}{a character object specifying which subject survival analysis
is applying to.}

\item{ptable.df}{output from function survivalByQuantile. It includes information about
survival under various threshold}

\item{alpha}{a numeric value representing the statistically significant p-value}
}
\value{
a numeric value representing area for each taxlev is stored in a dataframe
}
\description{
This function calculate the distence between 0.05 and p-value under 0.05 as an
indicator for area under the curve
}
