% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly_input.R
\name{butterfly_input}
\alias{butterfly_input}
\title{Generate input for butterfly plot}
\usage{
butterfly_input(area.input, subtype, num_micro)
}
\arguments{
\item{area.input}{a dataframe with area for all taxonomy level generated by 'calculateArea'}

\item{subtype}{a string object specifying what cancer subtype to be included}

\item{num_micro}{an integer indicating how many microbes shoul be included for zoomed up version}
}
\value{
a list with two ggplot2 objects. One for all taxonomy level and another for selected number speficied in 'num_micro'
}
\description{
Generate input for butterfly plot
}
