% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateArea.R
\name{calculateArea}
\alias{calculateArea}
\title{Quantifying taxonomy significance with multi-threshold object}
\usage{
calculateArea(taxlev, threshold.mat, alpha = 0.05)
}
\arguments{
\item{taxlev}{a character object specifying which subject survival analysis
is applying to.}

\item{threshold.mat}{output from function survivalByQuantile. It includes information about
survival under various threshold}

\item{alpha}{a numeric value representing the statistically significant p-value}
}
\value{
a numeric value representing area for each taxlev is stored in a dataframe
}
\description{
This function calculate the distence between 0.05 and p-value under 0.05 as an
indicator for area under the curve
}
