% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_threshold.R
\name{make_threshold}
\alias{make_threshold}
\title{Visulize Threshold vs. p-value}
\usage{
make_threshold(ptable.df, taxlev)
}
\arguments{
\item{ptable.df}{output from function survivalByQuantile. It includes information about
survival under various threshold}

\item{taxlev}{a character object specifying which subject survival analysis
is applying to.}
}
\value{
A ggplot
}
\description{
Visulize Threshold vs. p-value
}
