% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StackBarinput.R
\name{StackBarinput}
\alias{StackBarinput}
\title{Title}
\usage{
StackBarinput(TCGA.list, ORIEN.list)
}
\arguments{
\item{TCGA.list}{a list object containing survival and taxonomy information for TCGA}

\item{ORIEN.list}{a list object containing survival and taxonomy information for ORIEN}
}
\value{
a dataframe that can be used to make stacked bar plot
}
\description{
Title
}
