% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_plot.R
\name{survival_plot}
\alias{survival_plot}
\title{Make survival curve}
\usage{
survival_plot(modified_input, taxlev, title.input)
}
\arguments{
\item{modified_input}{the output list from 'generate_surv_input' function}

\item{taxlev}{an output string object from 'generate_surv_input' specifying a taxonomy level to be plotted}

\item{title.input}{a string object to speficy the title of the plot}
}
\value{
a ggplot object
}
\description{
Make survival curve
}
