% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToWide.R
\name{ToWide}
\alias{ToWide}
\title{change data frame into wide format and match survival data}
\usage{
ToWide(
  data.long,
  surv.dat,
  taxalevels = c("domain", "kingdom", "phylum", "class", "order", "family", "genus",
    "species")
)
}
\arguments{
\item{data.long}{input taxonomy matrix in long format}

\item{surv.dat}{survival information for the data.long}

\item{taxalevels}{character vector representing every taxonomy level to be changed into wide format}
}
\value{
a long format of the taxonomy data with matching survival information
}
\description{
change data frame into wide format and match survival data
}
\examples{


}
